# API_SPECIFICATION.yaml
openapi: 3.0.3
info:
  title: Seven Steps to Poem API
  description: AI Agent system for automated business problem solving using McKinsey's 7-step methodology
  version: 1.0.0
  contact:
    name: API Support
    email: api-support@sevenstepstopoem.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.sevenstepstopoem.com/v1
    description: Production server
  - url: https://staging-api.sevenstepstopoem.com/v1
    description: Staging server
  - url: http://localhost:8000/v1
    description: Development server
security:
  - BearerAuth: []
paths:
  /problems:
    post:
      summary: Create a new problem analysis request
      description: Submit a business problem for AI agent analysis
      tags:
        - Problems
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - raw_text
                - submitter
              properties:
                raw_text:
                  type: string
                  description: Natural language description of the business problem
                  example: "Our SaaS customer churn rate has increased 20% over the past 3 months. We need to reduce it by 5% within 90 days."
                submitter:
                  type: string
                  format: email
                  description: Email address of the person submitting the problem
                  example: "john.doe@company.com"
                attachments:
                  type: array
                  items:
                    type: string
                    format: binary
                  description: Supporting files (PDF, CSV, XLSX)
                urgency:
                  type: string
                  enum: [low, medium, high, critical]
                  default: medium
                  description: Priority level of the problem
                metadata:
                  type: object
                  description: Additional context and metadata
                  properties:
                    industry:
                      type: string
                      example: "SaaS"
                    company_size:
                      type: string
                      enum: [startup, small, medium, large, enterprise]
                    budget_range:
                      type: string
                      example: "$10K-50K"
      responses:
        '201':
          description: Problem created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemResponse'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
        '429':
          description: Rate limit exceeded

    get:
      summary: List all problems for the authenticated user
      description: Retrieve a paginated list of problems
      tags:
        - Problems
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: status
          in: query
          schema:
            type: string
            enum: [created, framing, tree_generation, prioritizing, planning, analyzing, synthesizing, presenting, completed, failed]
        - name: sort
          in: query
          schema:
            type: string
            enum: [created_at, updated_at, urgency]
            default: created_at
        - name: order
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: desc
      responses:
        '200':
          description: List of problems
          content:
            application/json:
              schema:
                type: object
                properties:
                  problems:
                    type: array
                    items:
                      $ref: '#/components/schemas/ProblemSummary'
                  pagination:
                    $ref: '#/components/schemas/PaginationInfo'
  /problems/{problem_id}:
    get:
      summary: Get problem details
      description: Retrieve detailed information about a specific problem
      tags:
        - Problems
      parameters:
        - name: problem_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Problem details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '404':
          description: Problem not found

    delete:
      summary: Delete a problem
      description: Delete a problem and all associated data
      tags:
        - Problems
      parameters:
        - name: problem_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Problem deleted successfully
        '404':
          description: Problem not found
...
  /problems/{problem_id}/run-step:
    post:
      summary: Execute analysis step
      description: Trigger execution of a specific analysis step
      tags:
        - Analysis Steps
      parameters:
        - name: problem_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - step
              properties:
                step:
                  type: string
                  enum: [frame, tree, prioritize, plan, analyze, synthesize, present]
                  description: The analysis step to execute
                options:
                  type: object
                  description: Step-specific options
                  properties:
                    force_refresh:
                      type: boolean
                      default: false
                      description: Force re-execution even if results exist
                    include_artifacts:
                      type: boolean
                      default: true
                      description: Include generated artifacts in response
      responses:
        '202':
          description: Step execution started
          content:
            application/json:
              schema:
                type: object
                properties:
                  execution_id:
                    type: string
                    format: uuid
                  status:
                    type: string
                    enum: [queued, running, completed, failed]
                  estimated_duration:
                    type: integer
                    description: Estimated duration in seconds
        '409':
          description: Step already in progress
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ProblemResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        status:
          type: string
          enum: [created, framing_pending]
        created_at:
          type: string
          format: date-time
        estimated_completion:
          type: string
          format: date-time
    ProblemSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        status:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        urgency:
          type: string
        progress_percentage:
          type: integer
          minimum: 0
          maximum: 100
    ProblemDetail:
      type: object
      properties:
        id:
          type: string
          format: uuid
        raw_text:
          type: string
        submitter:
          type: string
        status:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        metadata:
          type: object
        progress:
          type: object
          properties:
            current_step:
              type: string
            completed_steps:
              type: array
              items:
                type: string
            progress_percentage:
              type: integer
...
    ErrorResponse:
      type: object
      properties:
        error_code:
          type: string
        error_message:
          type: string
        details:
          type: object
        timestamp:
          type: string
          format: date-time
        request_id:
          type: string
          format: uuid
    PaginationInfo:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total_items:
          type: integer
        total_pages:
          type: integer
        has_next:
          type: boolean
        has_previous:
          type: boolean
...